Moving common items between SAML and Social into a 'base'
