+++ awx/sso/social_pipeline.py
@@ -17,37 +17,6 @@ from awx.sso.common import get_or_create_with_default_galaxy_cred
 logger = logging.getLogger('awx.sso.social_pipeline')
 
 
-class AuthNotFound(AuthException):
-    def __init__(self, backend, email_or_uid, *args, **kwargs):
-        self.email_or_uid = email_or_uid
-        super(AuthNotFound, self).__init__(backend, *args, **kwargs)
-
-    def __str__(self):
-        return _('An account cannot be found for {0}').format(self.email_or_uid)
-
-
-class AuthInactive(AuthException):
-    def __str__(self):
-        return _('Your account is inactive')
-
-
-def check_user_found_or_created(backend, details, user=None, *args, **kwargs):
-    if not user:
-        email_or_uid = details.get('email') or kwargs.get('email') or kwargs.get('uid') or '???'
-        raise AuthNotFound(backend, email_or_uid)
-
-
-def set_is_active_for_new_user(strategy, details, user=None, *args, **kwargs):
-    if kwargs.get('is_new', False):
-        details['is_active'] = True
-        return {'details': details}
-
-
-def prevent_inactive_login(backend, details, user=None, *args, **kwargs):
-    if user and not user.is_active:
-        raise AuthInactive(backend)
-
-
 def _update_m2m_from_expression(user, related, expr, remove=True):
     """
     Helper function to update m2m relationship based on user matching one or
